requirements.txt
# Scrapping
requests~=2.28.1
beautifulsoup4~=4.11.1
bs4~=0.0.1
icrawler~=0.6.6

# Dataframe
pandas~=1.5.1

# Machine Learning
open_clip_torch~=2.3.1
torch~=1.13.0

# App
streamlit~=1.14.0
geopy~=2.2.0
